Rename experiments to packages
