+++ .package-template/pyproject.toml
@@ -1,0 +1,18 @@
+[project]
+name = "package-template"
+version = "0.1.0"
+description = "Add your description here"
+readme = "README.md"
+dependencies = [
+    # needed for logging in AzureML
+    "azureml-mlflow==1.58.*",
+    "mlflow-skinny==2.18.*",
+    "pytz==2024.2",
+]
+
+[project.scripts]
+package-template = "package_template:main"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
