+++ .package-template/src/.amlignore
