+++ container.env
@@ -8,7 +8,7 @@ TENANT_ID="<azure-tenant-id>"
 AZUREML_WORKSPACE="<azure-ml-workspace-name>"
 # Azure Machine Learning resource group scripts point to
 AZUREML_RESOURCE_GROUP="<azure-ml-resource-group>"
-# Folder where the main code lives
-CODE_PATH=./experiments
-# Relative folder where runs are packaged before submitting to AzureML
+# Folder where packages live
+PKGS_PATH=./packages
+# Relative folder where runs are isolated before submitting to AzureML
 RUNS_PATH=./runs
