--- experiments/example_experiment/src/example_experiment/__init__.py
@@ -1,2 +1,0 @@
-def main() -> None:
-    print("Hello from example-experiment!")
