+++ packages/example/pyproject.toml
@@ -1,17 +1,17 @@
 [project]
-name = "example-experiment"
+name = "example"
 version = "0.1.0"
 description = "Add your description here"
 readme = "README.md"
-requires-python = ">=3.12"
 dependencies = [
+    # needed for logging in AzureML
     "azureml-mlflow==1.58.*",
     "mlflow-skinny==2.18.*",
     "pytz==2024.2",
 ]
 
 [project.scripts]
-example-experiment = "example_experiment:main"
+example = "example:main"
 
 [build-system]
 requires = ["hatchling"]
