+++ packages/example/src/.amlignore
