+++ packages/shared/pyproject.toml
@@ -1,9 +1,8 @@
 [project]
 name = "shared"
 version = "0.1.0"
-description = "Package shared across experiments"
+description = "Package shared across the workspace"
 readme = "README.md"
-requires-python = ">=3.12"
 dependencies = []
